requirements.txt
Flask==3.0.3
playwright==1.47.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1